created a singleton function for web3modal instance
